Refactor project structure: Move models and repositories to db directory
