+++ frontend/src/db/adapters/http/base-http-repository.ts
