+++ frontend/src/db/models/inventory/table.ts
